[MAKI-84] - Верстка экрана Select Location. Заменил Bottom Sheet на обычную UIView
