+++ maki-maki-ios/SelectLocation/SelectLocationViewController.swift
@@ -10,14 +10,92 @@ import SnapKit
 import YandexMapsMobile
 
 final class SelectLocationViewController: UIViewController {
+        
+    // MARK: - Map View UI
     
-    // MARK: - UI
+    private lazy var mapView: UIView = {
+        let view = UIView()
+        return view
+    }()
+    
     private lazy var yandexMap = YMKMapView()
     
-    // MARK: - lifecycle
+    // MARK: - Address selection UI
+    
+    private lazy var addressSelectionView: UIView = {
+        let view = UIView()
+        view.backgroundColor = AppColor.background.uiColor
+        view.layer.cornerRadius = 20
+        return view
+    }()
+    
+    private lazy var selectLocationLabel: UILabel = {
+        let label = UILabel()
+        label.text = "Select Location"
+        label.textColor = AppColor.heading.uiColor
+        label.font = AppFont.bold.s24()
+        return label
+    }()
+    
+    private lazy var yourLocationLabel: UILabel = {
+        let label = UILabel()
+        label.text = "YOUR LOCATION"
+        label.textColor = AppColor.paragraph.uiColor
+        label.font = AppFont.reqular.s12()
+        return label
+    }()
+    
+    private lazy var locationTextField = OneLineTextField()
+    
+    private lazy var saveAsLabel: UILabel = {
+        let label = UILabel()
+        label.text = "Save As"
+        label.textColor = AppColor.paragraph.uiColor
+        label.font = AppFont.reqular.s12()
+        return label
+    }()
+    
+    private lazy var choiceLocationSegmentedController: UISegmentedControl = {
+        let control = UISegmentedControl(first: "Home", second: "Work", third: "Other")
+        control.layer.cornerRadius = 12
+        control.layer.masksToBounds = true
+        control.clipsToBounds = true
+        return control
+    }()
+        
+    private lazy var saveAddressButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.setTitle("SAVE ADDRESS", for: .normal)
+        button.tintColor = AppColor.heading.uiColor
+        button.backgroundColor = AppColor.accent.uiColor
+        button.layer.cornerRadius = 14
+        return button
+    }()
+
+    // MARK: - Lifecycle
+    
     override func viewDidLoad() {
         super.viewDidLoad()
 
+        setupViews()
+        setupConstraints()
+        setupYandexMap()
+    }
+    
+    // MARK: - Setup Views
+    
+    private func setupViews() {
+        [mapView, addressSelectionView,
+         selectLocationLabel, yourLocationLabel, locationTextField,
+         saveAsLabel, choiceLocationSegmentedController, saveAddressButton].forEach {
+            view.addSubview($0)
+        }
+        mapView.addSubview(yandexMap)
+    }
+    
+    // MARK: - Setup Yandex Map
+    
+    private func setupYandexMap() {
         yandexMap.mapWindow.map.move(
                 with: YMKCameraPosition.init(
                     target: YMKPoint(
@@ -29,53 +107,54 @@ final class SelectLocationViewController: UIViewController {
                     tilt: 0),
                 animationType: YMKAnimation(type: YMKAnimationType.smooth, duration: 5),
                 cameraCallback: nil)
-        
-        setupViews()
-        setupConstraints()
-        setupConfigureSheet()
     }
     
-    // MARK: - Setup Views
-    private func setupViews() {
-        view.addSubview(yandexMap)
-    }
+    // MARK: - Setup Constrainrs
     
-    // MARK: - Constraints
     private func setupConstraints() {
-        yandexMap.snp.makeConstraints { make in
-              make.edges.equalToSuperview()
+        mapView.snp.makeConstraints { make in
+            make.top.leading.trailing.equalToSuperview()
+            make.bottom.equalTo(addressSelectionView.snp.top).offset(20)
         }
-    }
-    
-    // MARK: - Configure Sheet
-    private func setupConfigureSheet() {
-        let viewController = BottomSheetViewController()
-        let navigationViewController = UINavigationController(rootViewController: viewController)
         
-//        navVC.isModalInPresentation = true
-//        if let sheet = navVC.sheetPresentationController {
-//            sheet.detents = [.custom(resolver: { context in
-//                return 378
-//            }), .medium()]
-//            sheet.largestUndimmedDetentIdentifier = .medium
-//        }
-        
-        navigationViewController.isModalInPresentation = true
-        if #available(iOS 15.0, *) {
-            if let sheet = navigationViewController.sheetPresentationController {
-                if #available(iOS 16.0, *) {
-                    sheet.detents = [.custom(resolver: { _ in
-                        return 378
-                    }), .medium()]
-                } else {
-                    // Fallback on earlier versions
-                }
-                sheet.largestUndimmedDetentIdentifier = .medium
-            }
-        } else {
-            // Fallback on earlier versions
+        addressSelectionView.snp.makeConstraints { make in
+            make.bottom.leading.trailing.equalToSuperview()
+            make.height.equalTo(378)
         }
-
-        navigationController?.present(navigationViewController, animated: true)
+        
+        selectLocationLabel.snp.makeConstraints { make in
+            make.top.equalTo(addressSelectionView.snp.top).offset(27)
+            make.leading.equalTo(addressSelectionView.snp.leading).offset(16)
+        }
+        
+        yourLocationLabel.snp.makeConstraints { make in
+            make.top.equalTo(selectLocationLabel.snp.bottom).offset(26)
+            make.leading.equalTo(addressSelectionView.snp.leading).offset(16)
+        }
+        
+        locationTextField.snp.makeConstraints { make in
+            make.top.equalTo(yourLocationLabel.snp.bottom).offset(12)
+            make.leading.equalToSuperview().offset(16)
+            make.trailing.equalToSuperview().offset(-16)
+        }
+        
+        saveAsLabel.snp.makeConstraints { make in
+            make.top.equalTo(locationTextField.snp.bottom).offset(45)
+            make.leading.equalTo(addressSelectionView.snp.leading).offset(16)
+        }
+        
+        choiceLocationSegmentedController.snp.makeConstraints { make in
+            make.top.equalTo(saveAsLabel.snp.bottom).offset(12)
+            make.leading.equalToSuperview().offset(16)
+            make.trailing.equalToSuperview().offset(-16)
+            make.height.equalTo(44)
+        }
+        
+        saveAddressButton.snp.makeConstraints { make in
+            make.top.equalTo(choiceLocationSegmentedController.snp.bottom).offset(32)
+            make.leading.equalToSuperview().offset(16)
+            make.trailing.equalToSuperview().offset(-16)
+            make.height.equalTo(53)
+        }
     }
 }
